backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==10.2.0
easyocr==1.7.2
opencv-python==4.12.0.88
numpy==1.26.4